Create README.md
This is one the deep learning technique used  for training images by the machine.The dataset consists of nearly 5000 cat images and 5000 dog images and here is Image is coverted into a array by ImageDataGenerator. The accuracy of both training and testing is being found and also a new image is being tested.